% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprcomp.lib.R
\name{comprcomp}
\alias{comprcomp}
\title{Function comprcomp.}
\usage{
comprcomp(X, Y, center = TRUE, scale = FALSE, ncomp = 0, ...)
}
\arguments{
\item{X}{A matrix of data points (raw data).}

\item{Y}{A vector of classes (factor).}

\item{center}{A logical value indicating centering (per variable/per column).}

\item{scale}{A logical value indicating scaling to unit variance (per variable/per column).}

\item{ncomp}{The numberof components.
The default value is \code{0}, that means computing of all components.}

\item{...}{Arguments to be passed to \code{cpca} function.}
}
\value{
A object of class \code{comprcomp}.
}
\description{
This function is inspired by the base function \code{prcomp} for PCA and 
shares similarities in arguments (\code{center} and \code{scale} arguments),
and methods (e.g. \code{scores}).
}
\details{
The difference is that \code{comprcomp} computes CPCA by calling \code{cpca} function.

The advantage of using \code{comprcomp} instead of \code{cpca} function 
is that the input data are given the format observaions vs. variables,
and construction of covariance matrices, as well as pre-procesing routines,
are implemented inside of \code{comprcomp} function.
Consequently, projection of new data, plotting, etc are easier
(as easy as using \code{prcomp} function).

\code{comprcomp} function is a constructor for \code{comprcomp} class.
}

