% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprcomp.lib.R
\name{comprcompClass}
\alias{comprcompClass}
\alias{print.comprcomp}
\alias{scoreplot.comprcomp}
\alias{scores.comprcomp}
\alias{varcomp}
\alias{varcomp.comprcomp}
\alias{varplot}
\alias{varplot.comprcomp}
\title{S3 class comprcomp.}
\usage{
\method{print}{comprcomp}(x, ...)

\method{scores}{comprcomp}(object, X, Y, comp = 1:2, center, scale,
  grouping = FALSE, ...)

varcomp(object, ...)

\method{varcomp}{comprcomp}(object, X, Y, comp, center, scale,
  grouping = FALSE, prop = FALSE, perc = FALSE, ...)

varplot(object, ...)

\method{varplot}{comprcomp}(object, X, Y, comp = 1:2, facet = TRUE, ...)

\method{scoreplot}{comprcomp}(object, X, Y, comp = 1:2, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{An object of class \code{comprcomp}.}

\item{X}{A matrix of data points (raw data).}

\item{Y}{A vector of classes (factor).}

\item{comp}{A vector of components (indcices).}

\item{center}{A logical value indicating centering (per variable/per column).}

\item{scale}{A logical value indicating scaling to unit variance (per variable/per column).}

\item{grouping}{A logical value indicating doing PCA or CPCA (PCA per group).}
}
\description{
S3 class comprcomp.
}

