\docType{package}
\name{cpca}
\alias{cpca}
\alias{cpca-package}
\title{Package cpc.}
\description{
  Methods to perform Common Principal Component Analysis
  (CPCA).
}
\references{
  Kanaan-Izquierdo, S., Ziyatdinov, A., Massanet, R., &
  Perera, A. (2012).  Multiview approach to spectral
  clustering.  In 2012 Annual International Conference of
  the IEEE Engineering in Medicine and Biology Society (pp.
  1254–1257). IEEE. doi:10.1109/EMBC.2012.6346165
}

