% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpc.lib.R
\name{cpca}
\alias{cpca}
\title{Function cpca.}
\usage{
cpca(cov, ng, ncomp = 2, method = "base", ...)
}
\arguments{
\item{cov}{A list of covariance matrices for groups under study.}

\item{ng}{A vector with the number of samples per group.}

\item{ncomp}{The number of components.
The default value is \code{2}.}

\item{method}{A character with the name of method to be used.
The default and the only value is \code{"base"}.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{CPCAPower}.
}
\description{
Function cpca.
}

