% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classCPCAPower.R
\name{CPCAPowerClass}
\alias{CPCAPowerClass}
\alias{print.CPCAPower}
\title{S3 class CPCAPower.}
\usage{
\method{print}{CPCAPower}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{CPCAPower}.}

\item{digits}{The number of digits to be printed.}

\item{...}{Additional arguments.}
}
\description{
S3 class CPCAPower.
}

