% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.lib.R
\name{eigenPower}
\alias{eigenPower}
\title{Function eigenPower.}
\usage{
eigenPower(A, v0, tol = 1e-06, maxit = 1000, sparse = FALSE,
  sparseSymm = FALSE, verbose = 0)
}
\arguments{
\item{A}{A two-dimensional square matrix, either of \code{matrix} or \code{Matrix} class.}

\item{v0}{A numeric vector; the initial guess for eignevector.
If it is missing, a random vector is generated.}

\item{tol}{The tolerance threshold used to stop when reaching no improvement if estmiation of eigenvalue.
The default value is \code{1e-6}.}

\item{maxit}{The maximum number of iterations.
The default value is \code{1e4}.}

\item{sparse}{The boolean value, whether to convert the input \code{A} matrix to one of the \code{Matrix} classes.
The default value is \code{FALSE}.}

\item{sparseSymm}{The boolean value, whether to convert the input \code{A} matrix to one of the \code{Matrix} classes,
while trying to convert to a symmetric matrix type.
The default value is \code{FALSE}.}

\item{verbose}{The integer value indicating the verbose level.
The default value is \code{0}.}
}
\value{
A list several slots: \code{v} the first eigenvector; 
  \code{lambda} the first eigenvalue; etc.
}
\description{
The function implements the power algorithm for EVD.
}
\note{
This function is inspired by the post \url{http://blogs.sas.com/content/iml/2012/05/09/the-power-method.html}.
}
\examples{
A <- matrix(c(-261, 209, -49, 
  -530, 422, -98,
  -800, 631, -144),
  ncol = 3, nrow = 3, byrow = TRUE)
  
v0 <- c(1, 2, 3)

out <- eigenPower(A, v0, maxit = 40, verbose = 2)

out[c("lambda", "v")]
}

