% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.lib.R
\name{eigenPowerArmaParallel}
\alias{eigenPowerArmaParallel}
\title{Function eigenPowerArmaParallel.}
\usage{
eigenPowerArmaParallel(A, v0, tol = 1e-06, maxit = 1000, cores = -1,
  chunkSize = 1, verbose = 0)
}
\arguments{
\item{A}{A two-dimensional square matrix, either of \code{matrix} or \code{Matrix} class.}

\item{v0}{A numeric vector; the initial guess for eignevector.
If it is missing, a random vector is generated.}

\item{tol}{The tolerance threshold used to stop when reaching no improvement if estmiation of eigenvalue.
The default value is \code{1e-6}.}

\item{maxit}{The maximum number of iterations.
The default value is \code{1e4}.}

\item{cores}{The number of cores.
The default value is \code{-1}.
This argument is passed next to \code{RcppParallel::setThreadOptions(numThreads = cores)}.}

\item{chunkSize}{The minimal size of a chunk.
The default value is \code{1}.
This argument is passed next to a wrapper \code{eigenPower_Arma_Parallel}.}

\item{verbose}{The integer value indicating the verbose level.
The default value is \code{0}.}
}
\value{
A list several slots: \code{v} the first eigenvector; 
  \code{lambda} the first eigenvalue; etc.
}
\description{
The function implements the power algorithm for EVD using RcppARmadillo and RcppParallel.
}

