% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classEigenPower.R
\name{EigenPowerClass}
\alias{EigenPowerClass}
\alias{print.EigenPower}
\title{S3 class EigenPower.}
\usage{
\method{print}{EigenPower}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{EigenPower}.}

\item{digits}{The number of digits to be printed.}

\item{...}{Additional arguments.}
}
\description{
S3 class EigenPower.
}

