% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classEigenPower.R
\name{EigenPowerRcppClass}
\alias{EigenPowerRcppClass}
\alias{print.EigenPowerRcpp}
\title{S3 class EigenPowerRcpp.}
\usage{
\method{print}{EigenPowerRcpp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{EigenPowerRcpp}.}

\item{...}{Additional arguments.}
}
\description{
S3 class EigenPowerRcpp.
}

