% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.lib.R
\name{eigenPowerRcppParallel}
\alias{eigenPowerRcppParallel}
\title{Function eigenPowerRcppParallel.}
\usage{
eigenPowerRcppParallel(A, v0, tol = 1e-06, maxit = 1000, cores,
  verbose = 0)
}
\arguments{
\item{A}{A two-dimensional square matrix, either of \code{matrix} or \code{Matrix} class.}

\item{v0}{A numeric vector; the initial guess for eignevector.
If it is missing, a random vector is generated.}

\item{tol}{The tolerance threshold used to stop when reaching no improvement if estmiation of eigenvalue.
The default value is \code{1e-6}.}

\item{maxit}{The maximum number of iterations.
The default value is \code{1e4}.}

\item{cores}{The number of cores.
If it is specified, \code{RcppParallel::setThreadOptions(numThreads = cores)} is called.}

\item{verbose}{The integer value indicating the verbose level.
The default value is \code{0}.}
}
\value{
A list several slots: \code{v} the first eigenvector; 
  \code{lambda} the first eigenvalue; etc.
}
\description{
The function implements the power algorithm for EVD using RcppParallel.
}

