% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.lib.R
\name{hilbert}
\alias{hilbert}
\title{Function hilbert.}
\usage{
hilbert(n)
}
\arguments{
\item{n}{The matrix size.}
}
\value{
A symmetric matrix
}
\description{
This function is used to produce a symmetric matrix.
}
\references{
\url{https://stat.ethz.ch/R-manual/R-devel/library/base/html/svd.html}
}

