% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprcomp.lib.R
\name{comprcompClass}
\alias{comprcompClass}
\alias{print.comprcomp}
\alias{scoreplot}
\alias{scoreplot.comprcomp}
\alias{scores}
\alias{scores.comprcomp}
\alias{varcomp}
\alias{varcomp.comprcomp}
\alias{varplot}
\alias{varplot.comprcomp}
\title{S3 class comprcomp.}
\usage{
\method{print}{comprcomp}(x, ...)

scores(object, ...)

\method{scores}{comprcomp}(object, X, Y, comp = 1:2, center, scale,
  grouping = FALSE, ...)

varcomp(object, ...)

\method{varcomp}{comprcomp}(object, X, Y, comp, center, scale,
  grouping = FALSE, ...)

varplot(object, ...)

\method{varplot}{comprcomp}(object, X, Y, comp = 1:2, facet = TRUE, ...)

scoreplot(object, ...)

\method{scoreplot}{comprcomp}(object, X, Y, comp = 1:2, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{An object of class \code{comprcomp}.}
}
\description{
S3 class comprcomp.
}

