% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethodsS3.R, R/comprcomp.lib.R
\name{varplot}
\alias{comprcompClass}
\alias{compscore}
\alias{compscore.comprcomp}
\alias{compvar}
\alias{compvar.comprcomp}
\alias{print.comprcomp}
\alias{scoreplot.comprcomp}
\alias{scoreplot.lda}
\alias{scoreplot.prcomp}
\alias{scores.comprcomp}
\alias{scores.lda}
\alias{varplot}
\alias{varplot.comprcomp}
\title{S3 class comprcomp.}
\usage{
varplot(object, ...)

compvar(object, ...)

compscore(object, ...)

\method{print}{comprcomp}(x, ...)

\method{scores}{comprcomp}(object, X, Y, comp = 1:2, center, scale,
  grouping = FALSE, ...)

\method{scores}{lda}(object, X, Y, comp = 1:2, ...)

\method{compscore}{comprcomp}(object, X, sorted = FALSE, ...)

\method{compvar}{comprcomp}(object, X, Y, comp, center, scale,
  grouping = FALSE, prop = FALSE, perc = FALSE, sorted = FALSE, ...)

\method{varplot}{comprcomp}(object, X, Y, comp = 1:2, facet = TRUE, ...)

\method{scoreplot}{comprcomp}(object, X, Y, comp = 1:2, ...)

\method{scoreplot}{prcomp}(object, X, Y, comp = 1:2, ...)

\method{scoreplot}{lda}(object, X, Y, comp = 1:2, ...)
}
\arguments{
\item{object}{An object of class \code{comprcomp}.}

\item{...}{Additional arguments.}

\item{x}{An object of class \code{comprcomp}.}

\item{X}{A matrix of data points (raw data).}

\item{Y}{A vector of classes (factor).}

\item{comp}{A vector of components (indcices).}

\item{center}{A logical value indicating centering (per variable/per column).}

\item{scale}{A logical value indicating scaling to unit variance (per variable/per column).}

\item{grouping}{A logical value indicating doing PCA or CPCA (PCA per group).
The default value is \code{FALSE}.}

\item{sorted}{A logical indicating wether to sort results or not.
The default value is \code{FALSE}.}

\item{prop}{A logical indicating whether the data are to be reported in proportions.
The default value is \code{FALSE}.}

\item{perc}{A logical indicating whether the data are to be reported in percentage.
The default value is \code{FALSE}.}

\item{facet}{A logical indicating whether faceting is to applied in \code{varplot} method.}
}
\description{
S3 class comprcomp.
}

