% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comprcomp.lib.R
\name{comprcompClass}
\alias{comprcompClass}
\alias{scoreplot.comprcomp}
\alias{scores.comprcomp}
\alias{varplot}
\alias{varplot.comprcomp}
\title{S3 class comprcomp.}
\usage{
\method{scores}{comprcomp}(object, X, Y, comp = 1:2, center, scale,
  grouping = FALSE, ...)

varplot(object, ...)

\method{varplot}{comprcomp}(object, X, Y, comp = 1:2, facet = TRUE, ...)

\method{scoreplot}{comprcomp}(object, X, Y, comp = 1:2, ...)
}
\arguments{
\item{object}{An object of class \code{comprcomp}.}

\item{...}{Additional arguments.}
}
\description{
S3 class comprcomp.
}

