% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.lib.R
\name{eigenPowerRcpp}
\alias{eigenPowerRcpp}
\title{Function eigenPowerRcpp.}
\usage{
eigenPowerRcpp(A, v0, tol = 1e-06, maxit = 1000, mode = 1, verbose = 0)
}
\arguments{
\item{A}{A two-dimensional square matrix, either of \code{matrix} or \code{Matrix} class.}

\item{v0}{A numeric vector; the initial guess for eignevector.
If it is missing, a random vector is generated.}

\item{tol}{The tolerance threshold used to stop when reaching no improvement if estmiation of eigenvalue.
The default value is \code{1e-6}.}

\item{maxit}{The maximum number of iterations.
The default value is \code{1e4}.}

\item{mode}{The integer number encoding the mode of computation in `eigenPower_Rcpp` C++ function.
\code{1} means doing matrix operations in loops (doing some calculus sync.).
\code{2} means dogig matrix operations using functions like computing the vector norm, etc.}

\item{verbose}{The integer value indicating the verbose level.
The default value is \code{0}.}
}
\value{
A list several slots: \code{v} the first eigenvector; 
  \code{lambda} the first eigenvalue; etc.
}
\description{
The function implements the power algorithm for EVD using Rcpp.
}

